+++ backend/src/main/java/com/health_insurance/backend/repository/CoverPlanRepository.java
@@ -4,8 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.health_insurance.backend.model.CoverPlan;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 public interface CoverPlanRepository extends JpaRepository<CoverPlan, Long> {
-    Optional<CoverPlan> findByPersonaID(Long personaID);
+    Optional<CoverPlan> findByPersonaID(BigInteger personaID);
 }
